Moved resource_binding to internal
