+++ include/VkZero/Internal/bind_resource_internal.hpp
