+++ include/VkZero/image.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "VkZero/Internal/core_internal.hpp"
-#include "VkZero/bind_resource.hpp"
+#include "VkZero/Internal/bind_resource_internal.hpp"
 #include "VkZero/info.hpp"
 #include "VkZero/resource_manager.hpp"
 #include "VkZero/types.hpp"
